Move enum unwrappers to suitable AST nodes
If an unwrapper only calls methods from a single
AST node, it is placed in the same file
