+++ rust/parse_ast/src/ast_nodes/spread_element.rs
@@ -1,5 +1,5 @@
 use swc_common::Span;
-use swc_ecma_ast::Expr;
+use swc_ecma_ast::{Expr, SpreadElement};
 
 use crate::convert_ast::converter::ast_constants::{SPREAD_ELEMENT_ARGUMENT_OFFSET, SPREAD_ELEMENT_RESERVED_BYTES, TYPE_SPREAD_ELEMENT};
 use crate::convert_ast::converter::AstConverter;
@@ -21,5 +21,9 @@ impl<'a> AstConverter<'a> {
       .try_into()
       .unwrap();
     self.buffer[end_position..end_position + 4].copy_from_slice(&expression_end);
+  }
+
+  pub fn convert_spread_element(&mut self, spread_element: &SpreadElement) {
+    self.store_spread_element(&spread_element.dot3_token, &spread_element.expr);
   }
 }
